Feature/migrate to react on rails (#408)
* Add react_on_rails gem & package

* Init React on Rails

* Hide script tag

* Rewrite App entry point, add react on rails register for App and NavUserContainer

* Move component imports into startup file, add CoinShow to react on rails, move window size functions into application entry

* Add store wrapper, add IcoFilters to react on rails

* Add FlashMessageList to react on rails

* Add WatchButton to react on rails

* Add CalendarPage to react on rails

* Upgrade react_on_rails

* Move all react on rails register calls to `startup.js`

* Add comment for ReactOnRails.register

* Remove ReactOnRails.register from WatchButton

* Fix required build dependency

* Move all typings into 'dependencies'
