+++ Gemfile
@@ -34,6 +34,8 @@ gem 'redis-rails' # Remove when upgrading to Rails 5.2 since it is built-in
 # Use Capistrano for deployment
 # gem 'capistrano-rails', group: :development
 
+gem 'react_on_rails', '11.1.4'
+
 gem 'active_hash'
 gem 'acts-as-taggable-on'
 gem 'administrate'
@@ -105,3 +107,5 @@ end
 # Windows does not include zoneinfo files, so bundle the tzinfo-data gem
 # gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
 
+
+gem 'mini_racer', platforms: :ruby
