+++ Procfile.dev-server
@@ -1,0 +1,12 @@
+# You can run these commands in separate shells instead of using foreman
+web: rails s -p 3000
+
+# Next line runs the webpack-dev-server
+# You can edit config/webpacker.yml to set HMR to true to see hot reloading.
+# Note, hot and live reloading don't work with the default generator setup on top of
+# the rails/webpacker Webpack config with server rendering.
+# If you have server rendering enabled, modify the call to bin/webpack-dev-server line
+# so you add `--inline=false` and then CSS is not inlined.
+# Otherwise, you will have an error. If you want HMR and Server Rendering, see
+# the example in the https://github.com/shakacode/react-webpack-rails-tutorial
+client: sh -c 'rm -rf public/packs/* || true && bundle exec rake react_on_rails:locale && bin/webpack-dev-server'
