+++ app/assets/stylesheets/base/general.scss
@@ -1,3 +1,7 @@
+script {
+  display: none !important;
+}
+
 * {
   outline: none;
 }
