+++ app/javascript/App/components/IcoFilters.js
@@ -1,3 +1,5 @@
 import Filters from './Filters'
 import container from '../containers/icoFilters'
-export default container(Filters)
+import withStore from '../withStore'
+
+export default withStore(container(Filters))
