--- app/javascript/App/index.js
@@ -1,152 +1,0 @@
-/*
- * Normally this file would just mount the app on the document body, however
- * this is not an SPA, so it looks for <component> tags in the DOM and injects
- * React components there, optionally with the Redux store.
- */
-import 'babel-polyfill'
-
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { Provider } from 'react-redux'
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import configureStore from './configureStore'
-import getScreenSize from './lib/screenSize'
-import { PersistGate } from 'redux-persist/integration/react'
-import debounce from 'debounce'
-import { MuiThemeProvider } from '@material-ui/core/styles'
-import theme from './theme'
-
-import appContainer from './containers/app'
-
-import WatchButton from './components/WatchButton'
-import TwitterFeed from './components/TwitterFeed'
-import RedditFeed from './components/RedditFeed'
-import IcoFilters from './components/IcoFilters'
-import ExchangeListingsPage from './bundles/ExchangeListings/ExchangeListingsCointainer'
-import NewsfeedPageNew from './bundles/NewsfeedPage/NewsfeedPageContainer'
-import Tabs from './components/Tabs'
-import CoinCharts from './components/CoinCharts'
-import CalendarPage from './components/CalendarPage'
-import CoinIndex from './components/CoinIndex'
-import CoinShow from './components/CoinShow'
-import scrollHelper from './scrollHelper'
-import CoinListContainer from './bundles/common/containers/CoinListContainer'
-import NewsfeedContainer from './bundles/NewsfeedPage/NewsfeedContainer'
-import FlashMessageListContainer from './bundles/common/containers/FlashMessageListContainer'
-import NavUserContainer from './bundles/common/containers/NavUserContainer'
-
-const injectableComponents = {
-  WatchButton,
-  TwitterFeed,
-  RedditFeed,
-  IcoFilters,
-  Tabs,
-  CoinCharts,
-  CalendarPage,
-  CoinShow,
-  FlashMessageListContainer,
-  NavUserContainer,
-}
-
-const injectComponents = () => {
-  const hooks = document.getElementsByTagName('component')
-  if (hooks) {
-    const { store, persistor } = configureStore()
-    Array.from(hooks).forEach((hook) => {
-      const name = hook.getAttribute('name')
-      const props = JSON.parse(hook.getAttribute('props'))
-
-      if (name === 'App') {
-        ReactDOM.render(
-          <MuiThemeProvider>
-            <NewsfeedContainer>
-              <CoinListContainer loggedIn={!!props.user}>
-                <Router>
-                  <Switch>
-                    <Route
-                      exact
-                      path="/news/:coinSlug?"
-                      render={(routeProps) => (
-                        <NewsfeedPageNew
-                          loggedIn={!!props.user}
-                          coinSlug={routeProps.match.params.coinSlug}
-                          topCoinSlugs={props.topCoinSlugs}
-                          categories={props.categories}
-                          feedSources={props.feedSources}
-                        />
-                      )}
-                    />
-                    <Route
-                      exact
-                      path="/news/:newsItemId/:newsItemSlug"
-                      render={(routeProps) => (
-                        <NewsfeedPageNew
-                          loggedIn={!!props.user}
-                          newsItemId={routeProps.match.params.newsItemId}
-                          categories={props.categories}
-                          feedSources={props.feedSources}
-                        />
-                      )}
-                    />
-                    <Route
-                      exact
-                      path="/listings"
-                      render={() => <ExchangeListingsPage {...props} />}
-                    />
-                    <Route
-                      exact
-                      path="/coins"
-                      render={() => <CoinIndex {...props} />}
-                    />
-                  </Switch>
-                </Router>
-              </CoinListContainer>
-            </NewsfeedContainer>
-          </MuiThemeProvider>,
-          hook,
-        )
-      } else {
-        const Component = injectableComponents[name]
-        if (!Component) return console.error(`Component "${name}" not found`)
-        const withStore = hook.getAttribute('withStore') !== null
-        if (withStore) {
-          const AppComponent = appContainer(Component)
-          ReactDOM.render(
-            <MuiThemeProvider theme={theme}>
-              <CoinListContainer user={props.user}>
-                <Provider store={store}>
-                  <PersistGate loading={null} persistor={persistor}>
-                    <AppComponent {...props} />
-                  </PersistGate>
-                </Provider>
-              </CoinListContainer>
-            </MuiThemeProvider>,
-            hook,
-          )
-        } else {
-          ReactDOM.render(
-            <MuiThemeProvider theme={theme}>
-              <Component {...props} />
-            </MuiThemeProvider>,
-            hook,
-          )
-        }
-      }
-    })
-  }
-}
-
-const setScreenSize = () => {
-  window.screenSize = getScreenSize()
-  window.isMobile = !['m', 'l'].includes(window.screenSize)
-  window.isTablet = window.screenSize === 'm'
-  window.isDesktop = window.screenSize === 'l'
-}
-
-document.addEventListener('DOMContentLoaded', () => {
-  setScreenSize()
-  injectComponents()
-})
-
-window.addEventListener('resize', debounce(setScreenSize), 400)
-window.addEventListener('resize', debounce(scrollHelper), 400)
