+++ app/views/coins/show.slim
@@ -1,1 +1,1 @@
-.antd= react_component "CoinShow", withStore: true, props: { symbol: @coin.symbol, priceData: @coin.prices_data, annotations: @coin.news_data, isTradingViewVisible: true, coinObj: @coin, modeAnalyticsUrl: @mode_analytics_url, user: current_user, watching: user_is_watching?(@coin) }
+.antd= react_component "CoinShow", props: { symbol: @coin.symbol, priceData: @coin.prices_data, annotations: @coin.news_data, isTradingViewVisible: true, coinObj: @coin, modeAnalyticsUrl: @mode_analytics_url, user: current_user, watching: user_is_watching?(@coin) }
