+++ app/views/components/_flash.slim
@@ -1,3 +1,3 @@
 - if flash.any?
   .flex-none
-    = react_component "FlashMessageListContainer", withStore: false, props: { messages: flash_messages }
+    = react_component "FlashMessageListContainer", props: { messages: flash_messages }
