+++ app/views/icos/index.slim
@@ -7,6 +7,6 @@
       = render partial: 'nav'
   .bg-white.z-2.pa3.ph4-m
     .container-wide
-      = react_component 'IcoFilters', withStore: true, props: @react_props
+      = react_component 'IcoFilters', props: @react_props
       #ico-table.pv3
         = render partial: 'table'
