+++ app/views/pages/calendar.slim
@@ -1,1 +1,1 @@
-= react_component "CalendarPage", withStore: true, props: { categories: NewsCategory.pluck(:name), loggedIn: !!current_user }
+= react_component "CalendarPage", props: { categories: NewsCategory.pluck(:name), loggedIn: !!current_user }
