+++ package.json
@@ -12,10 +12,13 @@
     "@rails/webpacker": "^3.3.1",
     "@types/react": "^16.4.9",
     "@types/react-dom": "^16.0.7",
+    "@types/react-router-dom": "^4.3.0",
+    "@types/react-select": "^2.0.2",
     "antd": "^3.7.3",
     "axios": "^0.18.0",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.4",
+    "babel-plugin-root-import": "^6.1.0",
     "babel-plugin-styled-components": "^1.5.1",
     "babel-preset-react": "^6.24.1",
     "d3-ease": "^1.0.3",
@@ -41,6 +44,7 @@
     "react-dom": "^16.2.0",
     "react-infinite-scroll-component": "^4.2.0",
     "react-move": "^2.7.0",
+    "react-on-rails": "11.1.4",
     "react-onclickoutside": "^6.7.1",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.3.1",
@@ -77,11 +81,8 @@
     "@storybook/addon-links": "^3.4.8",
     "@storybook/addons": "^3.4.8",
     "@storybook/react": "^3.4.8",
-    "@types/react-router-dom": "^4.3.0",
-    "@types/react-select": "^2.0.2",
     "babel-eslint": "^8.1.1",
     "babel-plugin-import": "^1.8.0",
-    "babel-plugin-root-import": "^6.1.0",
     "babel-runtime": "^6.26.0",
     "core-js": "^2.5.7",
     "eslint": "^4.19.1",
